API Documentation

Table of Contents

API Documentation

Table of Contents

Index

Public Interface

Types

Functions

Index

MethodMoments.GMM

MethodMoments.Regression

Base.summary

MethodMoments.aic

MethodMoments.bic

MethodMoments.coef

MethodMoments.er

MethodMoments.er_std

MethodMoments.er_var

MethodMoments.gmm

MethodMoments.intercept

MethodMoments.jacobian

MethodMoments.llh

MethodMoments.momcov

MethodMoments.moments

MethodMoments.r2

MethodMoments.reg

MethodMoments.reoptimize

MethodMoments.vcov

MethodMoments.wald

Optim.optimize

Statistics.cor

Statistics.cov

Statistics.std

Statistics.var

StatsAPI.fit

StatsBase.sample

Public Interface

Types

MethodMoments.GMM — Type

Type GMM stores data related to the generalized method of moments corresponding to the condition E[f(θ)] = 0. The GMM estimator solves

Min g(θ)' W g(θ)

where g(θ) is the sample average of f(θ).

Use function gmm to define an instance of GMM, and then optimize to solve the minimization.

Fields

func::Function: sample function f

params::Vector: parameter vector θ

weight::Matrix: weighting matrix W

longcov::Matrix: estimate for long-run covariance matrix (see vcov)

twostep::Bool: true if GMM problem was reoptimized using efficient choice of W (see reoptimize)

nobs::Int: sample size

nmom::Int: number of moments

npar::Int: number of parameters

MethodMoments.Regression — Type

Type Regression stores data related to the estimation of the linear regression

Y = X β + e

through independent instruments Z. In the ordinary least squares (OLS) case, Z = X.

Use function reg to define an instance of Regression.

Fields

G::GMM: solution to GMM problem on corresponding to orthogonality conditions

Y::Vector: sample of dependent variable

X::Matrix: sample of right-hand side variables

Z::Matrix: sample of instrumental variables

W::Matrix: weighting matrix

intercept::Bool: true if model estimated with an intercept

nobs::Int: sample size

nins::Int: number of instrumental variables

nrhs::Int: number of right-hand side variables

Functions

MethodMoments.gmm — Function

gmm(f, θ, W=I)

Create a GMM object for the generalized method of moments problem Min g(θ)' W g(θ) where g(θ) is the sample average of f(θ). See GMM.

To solve the minimization, use optimize.

Arguments

f::Function: sample of moments function (output must be Vector or Matrix).

θ::Vector: parameter vector.

W::Matrix: weighting matrix (default = identity).

StatsBase.sample — Function

sample(G::GMM)

Compute sample of moments f(θ) (dim 1: observations, dim 2: moments). See documentation of type GMM for notation.

MethodMoments.moments — Function

moments(G::GMM)

Compute vector of empirical moments g(θ), which is the sample average of f(θ). See documentation of type GMM for notation.

MethodMoments.vcov — Function

vcov(G::GMM; lags=0)

Returns a new GMM object, updating the estimate of the long-run covariance matrix. The new estimate is the heteroskedasticity and autocorrelation consistent (HAC) Newey-West estimator with specified lags. When lag = 0, the Newey-West reduces to the Huber-White estimator.

vcov(R::Regresion; lags=0)

Returns a new Regression object, updating the estimate of the long-run covariance matrix. The new estimate is the heteroskedasticity and autocorrelation consistent (HAC) Newey-West estimator with specified lags. When lag = 0, the Newey-West reduces to the Huber-White estimator.

MethodMoments.jacobian — Function

jacobian(G::GMM)

Compute the jacobian matrix ∂g(θ)/∂θ by finite differences (dim 1: moments, dim 2: paramaters). See documentation of type GMM for notation.

Optim.optimize — Function

optimize(G::GMM; <kwargs>)

Solve the minimization problem Min g(θ)' W g(θ). See documentation of type GMM for notation.

The optimization uses the optimize function from the Optim package, passing over all keyword arguments (e.g. method=Optim.BFGS()).

optimize returns a GMM object storing the parameter vector θ that solves the minimization problem. The long-run covariance matrix is not replaced (use vcov).

To solve the minimization problem again using the feasible efficient weighting matrix, use reoptimize.

MethodMoments.reoptimize — Function

reoptimize(G::GMM; <kwargs>)

Solve the minimization problem Min g(θ)' S⁻¹ g(θ), where S is the estimated long-run covariance matrix stored in a. See documentation of type GMM for notation.

The optimization uses the optimize function from the Optim package, passing over all keyword arguments (e.g. method=Optim.BFGS()).

optimize returns a GMM object storing the parameter vector θ that solves the minimization problem. The long-run covariance matrix is not re-estimated (use vcov).

Statistics.cov — Function

cov(G::GMM)

Compute the covariance matrix of optimized parameter vector θ. See documentation of type GMM for notation.

cov(R::Regression)

Compute the covariance matrix of linear coefficients β, excluding the intercept term. Run cov(o.G) to include intercept. See documentation of type Regression for notation.

Statistics.var — Function

var(G::GMM)

Compute the (element-by-element) variance of optimized parameter vector θ. See documentation of type GMM for notation.

var(R::Regression)

Compute the (element-by-element) variance of linear coefficients β, excluding the intercept term. See documentation of type Regression for notation.

Statistics.std — Function

std(G::GMM)

Compute the (element-by-element) standard deviation of optimized parameter vector θ. See documentation of type GMM for notation.

var(R::Regression)

Compute the (element-by-element) standard deviation of linear coefficients β, excluding the intercept term. See documentation of type Regression for notation.

Statistics.cor — Function

cor(G::GMM)

Compute the correlation matrix of optimized parameter vector θ. See documentation of type GMM for notation.

cov(R::Regression)

Compute the correlation matrix of linear coefficients β, excluding the intercept term. See documentation of type Regression for notation.

MethodMoments.momcov — Function

momcov(G::GMM)

Compute the (singular) covariance matrix of minimized moments g(θ). See documentation of type GMM for notation.

Base.summary — Function

summary(G::GMM)

Print summary of the GMM problem.

summary(R::Regression)

Print summary of the linear regression.

MethodMoments.wald — Function

wald(G::GMM; R=I, r=0, subset)

Print result of the Wald test of the null R θ[subset] = r. See documentation of type GMM for notation.

If omitted, subset defaults to all parameters 1:G.npar. R defaults to the identity, and r defaults to a vector of zeros.

wald(re::Regression; R=I, r=0, subset)

Print result of the Wald test of the null R β[subset] = r, where β excludes the intercept. See documentation of type Regression for notation.

If omitted, subset defaults to all parameters (excluding the intercept). R defaults to the identity, and r defaults to a vector of zeros.

MethodMoments.reg — Function

reg(Y, X, Z, W; intercept)
reg(Y, X, Z)
reg(Y, X)

Create a Regression object with the instrumental variable (IV) estimator of β in the linear regression Y = X β + e.

If weighting matrix W is omitted, use identity.

If sample of instruments Z is ommitted, estimate β by ordinary least squares (OLS).

Arguments

Y::Vector: sample of dependent variable

X::Matrix: sample of right-hand side variables

Z::Matrix: sample of instrumental variables

W::Matrix: weighting matrix

intercept::Bool: true to add intercept on both X and Z (default: true)

MethodMoments.coef — Function

coef(R::Regression)

Return coefficient vector β, excluding intercept. To get intercept, use intercept. See documentation of type Regression for notation.

MethodMoments.intercept — Function

intercept(R::Regression)

Return intercept term.

StatsAPI.fit — Function

fit(R::Regression, X::VecOrMat)
fit(R::Regression)

Compute fit X β. See documentation of type Regression for notation. If X is not provided, use regressors stored in R (used for estimation).

MethodMoments.er — Function

er(R::Regression, Y::Vector, X::VecOrMat)
er(R::Regression)

Compute sample errors Y - X β. See documentation of type Regression for notation. If X and Y are not provided, use those stored in R (used for estimation).

MethodMoments.er_var — Function

er_var(R::Regression)

Return the estimated variance of the error term. No adjustment for degrees of freedom lost.

MethodMoments.er_std — Function

er_std(R::Regression)

Return the estimated standard deviation of the error term. No adjustment for degrees of freedom lost.

MethodMoments.r2 — Function

r2(R::Regression; adjust::Bool=false)

Return the coefficient of determination R², adjusted or not for model complexity.

MethodMoments.llh — Function

llh(R::Regression)

Return the log-likelihood of the estimated model.

MethodMoments.aic — Function

aic(R::Regression)

Return the Akaike Information Criterion.

MethodMoments.bic — Function

bic(R::Regression)

Return the Bayesian Information Criterion.